backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.5.0
motor==3.2.0
python-dotenv==1.0.0
python-multipart==0.0.6
watchdog==3.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
mongomock==4.3.0
mongomock-motor==0.0.34
bcrypt==4.2.1 
passlib==1.7.4
PyJWT==2.8.0